Move extract_panel_definitions_from_model_class and get_edit_handler to wagtail.admin.panels.model_utils
